Defining undefined fgColors for Python and others
For Python, I've set CLASSNAME and DEFNAME to `blue`/268BD2, i.e. the same as IDENTIFIER, as this appears to be in-line with what Solarized for vim does. TRIPLE and TRIPLEDOUBLE effectively serve as either constants or comments in Python depending on context, so there doesn't seem to be much consensus on styling. I've made the somewhat-arbitrary decision to use `magenta` for these as it's not used for anything else in Python currently.

All of the other undefined selectors I just changed to `base0`, as I don't have language samples or familiarity to test the rest. I'm fairly confident this will at least be an improvement on the default black in all cases.